requirements.txt
# RMSNormalization is supported in ONNX Opset 23 and later.
# We require onnx 1.18.0 and onnxruntime 1.20.0 or later for native support.
onnx>=1.18.0
onnxruntime-gpu>=1.20.0  # Use onnxruntime-gpu for TensorRT execution

# RMSNormalization is natively supported in TensorRT 10.15.1 and later.
# Ensure your system TensorRT library matches this version.
tensorrt>=10.15.1
